Change the lexer to store text in tokens rather than just the width
It seems like it might be a premature optimization to store only the
width. You would have to keep the source code around anyways if you
wanted to do anything with it. As long as we're not storing the position
of the token, it should be possible to do incremental reparsing.