Use pre-parser to handle indentation sensitivity
This is the same approach as F# takes; see the comment on the `preparse`
function in this commit. We run the token stream through a pre-parsing
phase, where we insert dummy tokens such as `in` to help the compiler
determine the end of constructs like `let`.

This supersedes the indent-token based approach, which probably would
not have worked anyways. The problem is that many indentation blocks may
be just for readability rather than to delineate a syntactic construct,
so we wouldn't know which `INDENT` tokens are significant and which
aren't.

This commit also converts `let` to be an expression and never a
statement, and only at the top level may this expression have an empty
expression after its `in`.