Parse 'let' expressions with some degree of error-recovery
This still needs to be exhaustively tested and all the corner-cases
worked out, by trying to remove every node that compromises the
let-expression and seeing it recover.

One interesting case would be recovering from something like this:

    let = "foo"

This would be a common case in the editor when the user is delete the
name of the variable on the right-hand side and replacing it with
something else. We should be able to just look ahead and see that it's
an '=' and then proceed with the parse, marking the right-hand side
pattern as missing.