Print out warnings in a nice way without crashing
It's likely that the user wants to attempt to use the compiler in a
best-effort way, even if we know that something is wrong. It's possible
that during a mis-parse we even recovered from an error in a sensible
way, and merely lost some of the tokens in the process.