Use `Range`s instead of `OffsetRange`s in diagnostics
We already take in an offset range and convert it into a range when
operating in phases that only have access to the green CST, such as the
parser. When operating in a phase that only hase access to the red CST,
then we only have regular positions, instead of offset ranges.

With the current architecture, we would have to convert a `Range` into
an `OffsetRange` and then back into a `Range`, which is a bit silly.
Instead of doing that, just pass in `Range`s directly and do the
conversion from `OffsetRange`s to `Range`s explicitly.