Add a green-to-red node type map
Consider the case of creating a new red CST node for the `n_expr` field
of the CST. We would write something like

    def n_expr(self) -> Expr:
        return Expr(
            ...
        )

However, we can't actually return an `Expr`, because this is an abstract
class. We must return a concrete subclass corresponding to the actual
type of the expression at runtime.

We could probably include some sort of `get_red_node` method into the
green node class, but this would introduce a circular dependency between
the green and red nodes (and circular dependencies probably won't be
supported in the language itself, which would make it harder to
migrate).

We could also add a long match case in the body of each function that
needs to construct a node of an abstract type, but it's easier to just
map the classes directly.