Use blue instead of green for notes
Since the errors are red and the notes are green, maybe people with
red-green colorblindness would have difficulty distinguishing them. It's
important to distinguish them because their ordering can be arbitrary:
the note may appear before or after in the file. So you can't simply
rely on the first or last message being the error message.