Update `.editorconfig` for `.pytch` files
I still have to decide what kind of spacing conventions to use. For now,
the mandate is two-space tabs.

Ideally, tabs would be used, since they carry the most semantic value,
and users can choose their desired tab size for display. But there are
some problems.

It's hard to tell what the size of a tab is. In the current error
display, we draw a box around source code, including on the right-hand
side. This can't be done unless we pick a size for the tab to be
displayed -- which sort of defeats the purpose of letting the user
select the tab size.

This is a little bit contrived, since we could produce error output in
away that doesn't have drawings on the right side (and arguably this
would be better because we can allow the terminal to wrap our text,
instead of doing it ourselves in a potentially-imperfect way).

A very real issue in this vein is autoformatting: how can you tell when
a line is too long when you can't even tell how long a line is? This is
only relevant if we decide to auto-break lines, but I'm leaning towards
doing that for now: coding styles are otherwise extremely inconsistent,
even withing code by the same author (in OCaml code, anyways).

Ruby is an example of a language which has effectively enforced
two-space tabs, so it can be done. Presumably some people don't like
that, but I don't think it's high on the list of reasons that people
don't use Ruby. Maybe it merits further digging on how much people don't
like it.

I think two-space tabs might be the standard for ML-family languages. I
should check.