Parser: Add parameter lists to let-expressions
This lets us make let bindings that are function definitions. So far,
type annotations for the parameter and return type aren't supported.

This is overly permissive in that you shouldn't be allowed to specify a
complex pattern along with a parameter list, such as this:

    let {foo}(x, y, z) = ...

This should be caught at a later stage in parsing, to simplify things.
(It seems that GHC does things this way, in that they parse in an
overly-permissive way and then may reject things later.)