Show diagnostic contexts in the order that they appear in the file
Rather than display the contexts of errors and notes in the order that
they were given, sort them such that they correspond to the ordering
they would be placed in the file. I'm finding that this leads to more
intuitive error displays; it's difficult to wrap one's head around the
relative location of a note accompanying an error message if it may be
logically before or after the error location in the file, but is always
rendered after in the error display.

I was thinking to have a rendering that uses a vertical colon, but I
couldn't figure out what to do with the message line; a blank line looks
awkward.