Add REPL
Closes #8. Unfortunately, this isn't a very good REPL: bindings created
in previous expressions aren't persisted through to new entries, and it
doesn't do tab-completion.

It's not trivial to make the bindings persist. We want to avoid
re-evaluating Pytch code that has already been written, in case it has
side effects (such as `print`), so we can't just append the new entry to
all of the old source code and re-run it.

We would have to use `exec` and its ability to pass in local variables,
which can then be modified by the user's entries. Naively, we could just
add all such local variables to the global scope for binding, but then
we might run into issues like this:

    >>> let foo = 1  # generates: `foo = 1`
    ...
    >>> let foo = 2  # generates: `foo2 = 2`
    ...
    >>> print(foo)   # `foo` refers to the wrong variable
    ...
    1
    >>> print(foo2)  # should produce binding error but doesn't
    ...
    2

This arises because the generated Python local variable names don't
match up with the Pytch binding names.