Type system: First implementation
This is a bidirectional type system based off of [Complete and Easy
Typechecking for Higher-Rank Polymorphism][Dunfield 2013].

  [Dunfield 2013]: https://www.cl.cam.ac.uk/~nk480/bidir.pdf

This is a simple but expressive type system. Some features:

  * It uses existential type variables rather than a constraint solver.
  I've heard that the constraint solver is one of the biggest pieces of
  production typecheckers, so avoiding it would be nice.
  * The paper sketches a way that it could subsume Hindley-Milner type
  inference, which would help make ML programmers comfortable with the
  type system.

It's unclear how some type system features will interact with this type
system (such as union and intersection types). Since the type system
does not have a constraint solver, and since I would like to keep it
that way, it's possible that these features will be check-able but not
infer-able, which is not a terrible downside.