Typechecker: implement "void" type for use in `if`-expressions
Python has no void type. Instead, it implicitly returns `None` from any
function that does not return a value. This is fine in Python, but this
implicit behavior can be a problem in Pytch, which is
expression-oriented.

Consider the following Pytch code:

    let foo =
      if cond()
      then 1
    print(foo)

What should be printed return if `cond()` returns `false`? There are
a few reasonable behaviors:

* It should print `None`.
* It should raise a `NameError` at runtime.
* It should be a compile-time error.

We choose the third. This is represented by requiring that
`if`-expressions without an `else` clause be of the "void" type.

Implications for the type system (denoting the void type as `Void` --
which is not the bottom type, as in Haskell):

* Noting that `None` is a singleton type, we say that `None <: Void`. A
value of `None` may be used anywhere that `Void` is expected, but not
the other way around. In particular, this enables
always-`None`-returning functions to be used in a context where `Void`
is expected. It does not enable, for example, the result of a function
returning `int | None` (that is, `Optional[int]`) to be assigned to
`Void`, as desired.
* `object` is no longer a top type, as we cannot assign a value of type
`Void` to `object`.

Now the above code will issue an error.

Future work:

* It might be that we should actually do most of the enforcement as a
lint rule, rather than directly in the typechecker.
  * We may not even want a void type, and instead just rely on the
  `None` type and the linter.
* Some of the error messages aren't terrible helpful. In particular,
it's not obvious to make an `if`-expression explicitly return `None` in
order to make it return `<void>`.
* Currently, you can instantiate an existential type variable to the
void type. Is that desirable?