Typesystem: add reasons
"Reasons" allow us to track why a type has a certain type. This can be
useful for debugging. In particular, we will want to show some of the
reasoning in error messages.

It's possible that I am tracking reasons too granularly for now, and
that we should only store reasons when it's possible that we will need
them for error messages, for efficiency purposes. (It is also possible
that there should be a "fast" typechecker mode where we don't allocate
reasons at all.)