Docs: Add language reference
This includes documentation for the syntax and semantics of all of the
language features that have been implemented so far (I believe).

The documentation around indentation is not great currently, as it's not
exhaustive. Expressions like `let`-expressions and `if`-expressions
don't have explicit documentation about how they interact with the
indentation stack in the preparser.

It doesn't currently include any details about the type system.