Fix bug with underlining one-character lines
We try to use a special character for pointing out one-character ranges
(currently an arrow pointing upward).

But I failed to consider the case where we want to underline multiple
lines and one of those lines just happens to have one character. In that
case, we should use the regular underline character, of course.