Check for unbuilt changes in CI
It's possible that there are unbuilt changes (or incorrectly-resolved
merge conflicts) in artifacts such as the documentation. We should
attempt to detect these in continuous integration. I don't want to add
this to the pre-commit hook because it's slow to rebuild documentation
from scratch, and the pre-commit hook should be as fast as possible.