Make pre-commit hook faster
The four steps in the pre-commit hook leave the repository unchanged, so
they can be run in parallel without issue. The only problem is that the
output is no longer intelligible or deterministic, since it may be
interleaved. We work around this by assuming that the pre-commit hook
passes most of the time, and only showing detailed output when it fails,
by re-running the pre-commit hook serially.