Bugfix: `return 1` from parallel pre-commit hook
Even though the exit code of `wait` is `1` and `set -e` is set, it
doesn't cause the function to return `1`. It seems to be because it's
ultimately inside a `!` expression in an `if`-statement:
https://stackoverflow.com/questions/4072984/how-do-i-get-effect-and-usefuless-of-set-e-inside-a-shell-function

It makes sense because you don't want the entire shell to exit in this
case, although it would be nice if `set -e` caused the function to
return `1` rather than attempt an `exit 1` under some circumstances.

This issue was caught by Travis CI's pre-commit hook, which I set to
always run in serial mode to catch issues like this.