Lexer: Support comments (closes #26)
Comments are the same as in Python, starting with `#`, and only having
the single-line variety.

There are some questions about how to lex multiline comments, in the
form

    # foo
    # bar
    baz

Logically, we may want to put them all in one comment trivium for ease
of tooling. But I think it will work out better to store their raw
representation and then rely on some higher-level functions to work with
them.

This commit also changes how we lex whitespace at the end of the file.
Previously, we would attach newlines as trailing trivia to the end of
the file, which is not a strictly wrong interpretation. But in this
commit, we incidentally change it to lex them as leading trivia for the
end of the file instead, which seems more harmonious, so that then there
really is no content following the EOF token.