Docs: Add CNAME support for Github Pages
A special `CNAME` file in the root of the `docs` directory is used to
link a Github Pages site to a real (non-`github.io`) domain.

The built-in Github Pages extension for Sphinx doesn't have support for
this, so we add this functionality with an extension of our own.
(Perhaps the built-in extension should have this functionality?)