Docs: Generate a sitemap
This might be useful for SEO purposes, as Google can process sitemaps.

In particular, I searched for "Pytch language" the other day, and the
language reference came before the homepage. I'm hoping that putting the
website behind a proper domain name and adding a sitemap will help
Google figure out that the homepage is more important.