Docs: Add changelog
The changelog is to be used to document detailed changes between
releases. This is especially useful for a still-under-development
programming language, since one may be waiting on some basic features to
be added, which can't be taken for granted.

Updating the changelog should be a part of most changes to the project.