Typechecker: Flag passing the wrong number of arguments
There are a few choices of syntax that we could flag:

* The entire function call.
* The parameters, including the parentheses.
* The parameters, not including the parentheses.
* Only extra parameters/where missing parameters would go.

I chose to select the parameters, including the parentheses. Flagging
the entire function call would be too noisy. Flagging only extra
parameters would likely be misleading, as it's not necessarily the case
that the extra parameters are at the end.

I figured it's best to include the parentheses to make it clear where
the arguments start and stop, but this could change.

(Note that in the case of flagging zero arguments, we would want to flag
the parentheses regardless, but this is not a blocker for the above, as
we could special-case that.)

In the future, we could also try to provide more detail about the
arguments that we expected. We could list out the arguments defined for
the function call and their types.