Parser, typechecker, codegen (#29): revise function definition syntax
What an ordeal! The details are at https://github.com/pytch-lang/pytch/issues/29.

The good news is that this also simplifies a lot of the code.
Previously, we used to check for parameter lists in the syntax tree and
do substantially different tasks in pretty much all phases of the
compiler. Now we use actually different nodes for these
logically-different constructs.